Revert "Merge branch 'main' into yakov.shapiro/send-mutate-event"
This reverts commit 6385e90954040b7e68e8aad1adb3a08bcc69411d, reversing
changes made to b721c29f7d214f6a04ccd72087a99e3433dea53d.

+++ pkg/network/event_windows.go
@@ -114,7 +114,7 @@ func FlowToConnStat(cs *ConnectionStats, flow *driver.PerFlowData, enableMonoton
 	cs.Family = family
 	cs.Direction = connDirection(flow.Flags)
 	cs.SPortIsEphemeral = IsPortInEphemeralRange(cs.Family, cs.Type, cs.SPort)
-	cs.Cookie = flow.FlowHandle
+	cs.Cookie = uint32(flow.FlowHandle)
 	if connectionType == TCP {
 		tf := flow.TCPFlow()
 		if tf != nil {
